// lib/supabase/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5";
  };
  public: {
    Tables: {
      games: {
        Row: {
          category: string;
          color: string;
          cover: string;
          created_at: string;
          description: string;
          id: string;
          is_published: boolean;
          sort_order: number;
          tagline: string;
          title: string;
        };
        Insert: {
          category: string;
          color: string;
          cover: string;
          created_at?: string;
          description: string;
          id: string;
          is_published?: boolean;
          sort_order: number;
          tagline: string;
          title: string;
        };
        Update: {
          category?: string;
          color?: string;
          cover?: string;
          created_at?: string;
          description?: string;
          id?: string;
          is_published?: boolean;
          sort_order?: number;
          tagline?: string;
          title?: string;
        };
        Relationships: [];
      };
      scores: {
        Row: {
          created_at: string;
          game_id: string;
          id: number;
          player_name: string;
          score: number;
        };
        Insert: {
          created_at?: string;
          game_id: string;
          id?: never;
          player_name: string;
          score: number;
        };
        Update: {
          created_at?: string;
          game_id?: string;
          id?: never;
          player_name?: string;
          score?: number;
        };
        Relationships: [
          {
            foreignKeyName: "scores_game_id_fkey";
            columns: ["game_id"];
            isOneToOne: false;
            referencedRelation: "game_stats";
            referencedColumns: ["game_id"];
          },
          {
            foreignKeyName: "scores_game_id_fkey";
            columns: ["game_id"];
            isOneToOne: false;
            referencedRelation: "games";
            referencedColumns: ["id"];
          },
        ];
      };
    };
    Views: {
      game_champions: {
        Row: {
          created_at: string | null;
          game_id: string | null;
          player_name: string | null;
          score: number | null;
        };
        Relationships: [
          {
            foreignKeyName: "scores_game_id_fkey";
            columns: ["game_id"];
            isOneToOne: false;
            referencedRelation: "game_stats";
            referencedColumns: ["game_id"];
          },
          {
            foreignKeyName: "scores_game_id_fkey";
            columns: ["game_id"];
            isOneToOne: false;
            referencedRelation: "games";
            referencedColumns: ["id"];
          },
        ];
      };
      game_stats: {
        Row: {
          best: number | null;
          game_id: string | null;
          plays: number | null;
        };
        Relationships: [];
      };
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  "public"
>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    keyof DefaultSchema["Enums"] | { schema: keyof DatabaseWithoutInternals },
  EnumName extends (DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never) = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never;

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends (PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never) = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never;